Bis zur Zusammenfassung Ok , Zusammenfassungs View muss noch komplettiert werden. Zum Ändern Rücksprung fehlt. Kontaktdatenerfassung fehlt.

+++ DeutscheBank.web/Models/ZusammenfassungModel.cs
@@ -13,7 +13,7 @@ namespace DeutscheBank.web.Models
             #endregion
 
             #region KreditRahmen
-            public int GewünschterBetrag { get; set; }
+            public double GewünschterBetrag { get; set; }
             public int Laufzeit { get; set; }
             #endregion
